refactor: implemento useYukigo hook
