+++ package.json
@@ -26,6 +26,7 @@
     "tailwindcss": "3",
     "web-vitals": "^2.1.4",
     "yukigo": "^0.2.3",
+    "yukigo-ast": "^0.2.1",
     "yukigo-haskell-parser": "^0.1.3"
   },
   "scripts": {
